Initial Push
** Work in Progress **

This Fork changes the following:

Add ASync non-blocking methods

Add Send and Recieve Timeouts

Move to .NET 6 for CancellationTokenSupport on UDPClient Send/Recv Methods

Rework implementation of retries to work with Send/Recieve timeouts

Fix issues with A2S_Info/ServerInfo properties like Port/SteamID//SourceTv/Keywords/GameID which rely on the EDF (Extra Data Flag) which was improperly detecting those other fields being supported

Obsolete Connect method, use Setup instead. UDP is connection-less.

*** Note on Retries and Send/Recieve Timeouts ***

Keep in mind the potentially worst case for execution time timeout*retries*4 (1 send auth, 1 receive auth, 1 send challenge, 1 receive result).

For example, if you have send/recv timeout set to 500ms, and timeout as default (10 times), and the server is dead/won't respond, it will take 5000ms to detect that, as the library will send 10 challenge packets which will timeout after 500ms.

+++ src/Okolni.Source.Query/Responses/InfoResponse.cs
@@ -113,7 +113,7 @@ namespace Okolni.Source.Query.Responses
         /// <summary>
         /// if ( EDF & 0x10 ) proves true: Server's SteamID.
         /// </summary>
-        public uint? SteamID { get; set; }
+        public long? SteamID { get; set; }
 
         /// <summary>
         /// if ( EDF & 0x40 ) proves true: Spectator port number for SourceTV.
@@ -135,7 +135,7 @@ namespace Okolni.Source.Query.Responses
         /// If this is present, a more accurate AppID is present in the low 24 bits. 
         /// The earlier AppID could have been truncated as it was forced into 16-bit storage.
         /// </summary>
-        public uint? GameID { get; set; }
+        public long? GameID { get; set; }
 
         /// <summary>
         /// If the Server is a The Ship Server
